Use the StorageProvider to set File#url; improve specs.
Although a file has a URL, it's not a property of the file itself.  (For
example, it should not be stored in the database, as then all records
would require rewriting if routing conventions change.)

The storage provider is well positioned to interrogate the routing layer
and decorate the file with a corresponding URL.  We will support Rails
routing conventions by default, but permit the user to override.

Improve the specs: configure the spec Uploader with our storage provider
to ensure correct method proxying; refactor.